requirements.txt
-e .
colour-runner==0.0.4
coverage==4.0.3
dj-inmemorystorage==1.4.0
dj-database-url==0.4.0
factory-boy==2.6.0
flake8==2.5.4
flake8-import-order==0.5.3
incuna-test-utils==6.4.1
psycopg2==2.6.1
